requirements.txt
Django==3.2.18
djangorestframework==3.14.0
markdown==django-filter3.5.2
django-filter==23.5
djangorestframework-simplejwt==5.3.1